🔨 Add useInterval Hook, Refactor Live, SSEDownNotice
Implemented request to have all new files be TypeScript.

Added a useInterval Hook (demo here: https://codesandbox.io/s/sessiontimer-useinterval-hook-demo-0l6jq)

Refactored Live workspace item and it's dependents to use hooks.

Refactored SSEDownNotice to use hooks.

Instead of useContext(Cerebral), now using useSignals() and useStore()
