--- packages/app/src/app/hooks/index.js
@@ -1,1 +1,0 @@
-export { useScript } from './useScript';
