+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/ConnectionNotice/ConnectionNotice.tsx
@@ -1,13 +1,13 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { observer } from 'mobx-react-lite';
-import { Cerebral } from 'app/store';
+import { useStore } from 'app/store';
 import { Container } from './elements';
 
 export const ConnectionNotice = observer(() => {
-  const { store } = useContext(Cerebral);
+  const { connected } = useStore();
 
   return (
-    !store.connected && (
+    !connected && (
       <Container>
         You{"'"}re not connected to the internet. You can still edit, but you
         cannot save. We recommend using the {"'"}Download{"'"} function to keep
