+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/SSEDownNotice/index.ts
@@ -1,0 +1,1 @@
+export { SSEDownNotice } from './SSEDownNotice';
