+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/User/index.ts
@@ -1,0 +1,1 @@
+export { User } from './User';
